rename a cpp file

+++ include/game/unknown.h
@@ -1,11 +1,11 @@
 #include "dolphin/types.h"
 u16 func_8002EC64();
-u32* lbl_803490E8;
-u32* func_8002EC5C();
-u32* lbl_803490E4;
-u32* func_8002EC54();
-u32* lbl_803490E0;
-u32* func_8002EC4C();
+u8* lbl_803490E8;
+u8* func_8002EC5C();
+u8* lbl_803490E4;
+u8* func_8002EC54();
+u8* lbl_803490E0;
+u8* func_8002EC4C();
 float FLOAT_8034a14c;
 float FLOAT_8034a148;
 double func_8002EC24(int param);
